Usando versão do Service Document e dos assets do instalador compatíveis com a branch da build

+++ api-src/app/service/build_service.py
@@ -4,6 +4,7 @@ import time
 import stat
 
 from app.settings import NEXUS_USER, NEXUS_PASSWORD, DOCKER_HUB_USER, DOCKER_HUB_PASSWORD, logger
+from app.service.service_document_downloader import ServiceDocumentDownloader
 
 
 class BuildService:
@@ -15,6 +16,8 @@ class BuildService:
         self.docker_hub_password = DOCKER_HUB_PASSWORD
         self.project_root_path = os.path.abspath(".\..")
         self.working_dir = os.path.join(self.project_root_path, f"temp\\{int(time.time())}")
+        self.erp_instalador_submodule_dir = os.path.join(self.working_dir, "erp-instalador")
+        self.service_document_downloader = ServiceDocumentDownloader()
 
     def execute(self, branch):
         # Restart scheduled shutdown (if set) to set a "timeout" in case anything goes wrong
@@ -28,7 +31,7 @@ class BuildService:
 
         # Download service document executable
         logger.info("Download Service Document Executable")
-        self.download_service_document()
+        downloaded_version = self.download_service_document(branch)
 
         # Build and upload image
         image_name = self.build_image(branch)
@@ -39,23 +42,32 @@ class BuildService:
         logger.info("Cleaning temporary files")
         self.clear_temporary_path()
 
-        return image_name
+        return {"image_name": image_name, "service_document_version": downloaded_version}
     
     def checkout_branch(self, branch):
-        os.system(f"git clone git@github.com:Nasajon/service-document-worker.git {self.working_dir} --recursive")
+        os.system(f"mkdir {self.working_dir}")
+        clone_response = os.system(f"git clone git@github.com:Nasajon/service-document-worker.git {self.working_dir} --recursive")
+        if clone_response != 0:
+            raise Exception("Exception when trying to clone worker repository")
         os.system(f"git checkout {branch}")
+        # Updates erp-instalador submodule
+        os.chdir(self.erp_instalador_submodule_dir)
+        os.system(f"git checkout {branch}")
+        os.system("git pull")
 
-    def download_service_document(self):
-        os.chdir(self.working_dir)
-        # TODO: baixar uma versão do service document equivalente a da branch sendo construída
-        os.system(f"curl.exe -u  {self.nexus_user}:{self.nexus_password} -X GET https://nexus.nasajon.com.br/repository/erp/br/com/nasajon/nsjServiceDocumentEngine/2.2203.0.1786/nsjServiceDocumentEngine-2.2203.0.1786.exe --output nsjServiceDocumentEngine.exe")
+    def download_service_document(self, branch):
+        download_abs_path = os.path.join(self.working_dir, "nsjServiceDocumentEngine.exe")
+        downloaded_version = self.service_document_downloader.download_latest_version_from_branch_name(branch, download_abs_path)
+        return downloaded_version
 
     def build_image(self, branch):
         os.chdir(self.working_dir)
         version = str(time.localtime().tm_mday) + str(time.localtime().tm_hour) + str(time.localtime().tm_min)
         image_name = f"arquiteturansj/servicedocumentworker:{branch}-{version}"
         logger.info(f"Building image {image_name}")
-        os.system(f"docker build -t {image_name} .")
+        build_response = os.system(f"docker build -t {image_name} .")
+        if build_response != 0:
+            raise Exception("Exception when trying to build docker image")
         return image_name
 
     def upload_image(self, image_name):
@@ -64,7 +76,9 @@ class BuildService:
         os.system(f"docker login -u {self.docker_hub_user} -p {self.docker_hub_password}")
 
         logger.info(f"Pushing image {image_name} to docker hub.")
-        os.system(f"docker push {image_name}")
+        push_response = os.system(f"docker push {image_name}")
+        if push_response != 0:
+            raise Exception("Exception when trying to push docker image to docker hub")
 
         logger.info(f"Deleting image {image_name} locally.")
         os.system(f"docker rmi {image_name}")
@@ -80,4 +94,7 @@ class BuildService:
         def remove_readonly(func, path, excinfo):
             os.chmod(path, stat.S_IWRITE)
             func(path)
-        shutil.rmtree(self.working_dir, onerror=remove_readonly)
+        try:
+            shutil.rmtree(self.working_dir, onerror=remove_readonly)
+        except Exception as e:
+            logger.exception(str(e))
